server: added authorization for retreiving only owned data

+++ server/src/main/java/me/vilius/homesync/controller/RoomController.java
@@ -8,13 +8,17 @@ import io.swagger.v3.oas.annotations.responses.ApiResponse;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import jakarta.persistence.EntityNotFoundException;
 import jakarta.validation.Valid;
+import me.vilius.homesync.model.CustomUserDetails;
 import me.vilius.homesync.model.Device;
 import me.vilius.homesync.model.Room;
+import me.vilius.homesync.model.User;
 import me.vilius.homesync.model.dto.RoomDTO;
+import me.vilius.homesync.service.HomeService;
 import me.vilius.homesync.service.RoomService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.Authentication;
 import org.springframework.validation.FieldError;
 import org.springframework.web.bind.MethodArgumentNotValidException;
 import org.springframework.web.bind.annotation.*;
@@ -30,6 +34,9 @@ public class RoomController extends BaseController {
     @Autowired
     private RoomService roomService;
 
+    @Autowired
+    private HomeService homeService;
+
     @Operation(summary = "Create a new room", description = "Creates a new room in a specific home")
     @ApiResponse(responseCode = "201", description = "Room created successfully",
             content = @Content(schema = @Schema(implementation = Room.class)))
@@ -39,7 +46,13 @@ public class RoomController extends BaseController {
     @PostMapping
     public ResponseEntity<?> createRoom(
             @Parameter(description = "ID of the home to add the room to") @RequestParam Long homeId,
-            @Parameter(description = "Room details") @Valid @RequestBody RoomDTO roomDTO) {
+            @Parameter(description = "Room details") @Valid @RequestBody RoomDTO roomDTO, Authentication authentication) {
+
+        User user = ((CustomUserDetails) authentication.getPrincipal()).getUser();
+        if (!homeService.userOwnsHome(user.getId(), homeId)) {
+            return new ResponseEntity<>("Unauthorized access to home", HttpStatus.FORBIDDEN);
+        }
+
         try {
             Room createdRoom = roomService.createRoom(homeId, roomDTO);
             return new ResponseEntity<>(createdRoom, HttpStatus.CREATED);
@@ -55,8 +68,10 @@ public class RoomController extends BaseController {
     @ApiResponse(responseCode = "200", description = "Successful retrieval of rooms",
             content = @Content(schema = @Schema(implementation = Room.class)))
     @GetMapping
-    public ResponseEntity<List<Room>> getAllRooms() {
-        return new ResponseEntity<>(roomService.getAllRooms(), HttpStatus.OK);
+    public ResponseEntity<List<Room>> getAllRooms(Authentication authentication) {
+        User user = ((CustomUserDetails) authentication.getPrincipal()).getUser();
+        List<Room> userRooms = roomService.getRoomsByUserHomes(user.getId());
+        return new ResponseEntity<>(userRooms, HttpStatus.OK);
     }
 
     @Operation(summary = "Get a room by ID", description = "Retrieves a room by its ID")
@@ -64,11 +79,16 @@ public class RoomController extends BaseController {
             content = @Content(schema = @Schema(implementation = Room.class)))
     @ApiResponse(responseCode = "404", description = "Room not found")
     @GetMapping("/{id}")
-    public ResponseEntity<?> getRoomById(@Parameter(description = "ID of the room to retrieve") @PathVariable Long id) {
+    public ResponseEntity<?> getRoomById(@Parameter(description = "ID of the room to retrieve") @PathVariable Long id
+            , Authentication authentication) {
+        User user = ((CustomUserDetails) authentication.getPrincipal()).getUser();
         Room room;
-        try{
+        try {
             room = roomService.getRoomById(id);
-        }catch (EntityNotFoundException e){
+            if (!homeService.userOwnsHome(user.getId(), room.getHome().getId())) {
+                return new ResponseEntity<>("Unauthorized access to room", HttpStatus.FORBIDDEN);
+            }
+        } catch (EntityNotFoundException e) {
             return new ResponseEntity<>("Room not found", HttpStatus.NOT_FOUND);
         }
         return new ResponseEntity<>(room, HttpStatus.OK);
@@ -82,8 +102,14 @@ public class RoomController extends BaseController {
     @PutMapping("/{id}")
     public ResponseEntity<?> updateRoom(
             @Parameter(description = "ID of the room to update") @PathVariable Long id,
-            @Parameter(description = "Updated room details") @Valid @RequestBody RoomDTO roomDTO) {
+            @Parameter(description = "Updated room details") @Valid @RequestBody RoomDTO roomDTO,
+            Authentication authentication) {
+        User user = ((CustomUserDetails) authentication.getPrincipal()).getUser();
         try {
+            Room room = roomService.getRoomById(id);
+            if (!homeService.userOwnsHome(user.getId(), room.getHome().getId())) {
+                return new ResponseEntity<>("Unauthorized access to room", HttpStatus.FORBIDDEN);
+            }
             Room updatedRoom = roomService.updateRoom(id, roomDTO);
             return new ResponseEntity<>(updatedRoom, HttpStatus.OK);
         } catch (EntityNotFoundException e) {
@@ -96,8 +122,14 @@ public class RoomController extends BaseController {
     @ApiResponse(responseCode = "204", description = "Room deleted successfully")
     @ApiResponse(responseCode = "404", description = "Room not found")
     @DeleteMapping("/{id}")
-    public ResponseEntity<?> deleteRoom(@Parameter(description = "ID of the room to delete") @PathVariable Long id) {
+    public ResponseEntity<?> deleteRoom(@Parameter(description = "ID of the room to delete") @PathVariable Long id,
+                                        Authentication authentication) {
+        User user = ((CustomUserDetails) authentication.getPrincipal()).getUser();
         try {
+            Room room = roomService.getRoomById(id);
+            if (!homeService.userOwnsHome(user.getId(), room.getHome().getId())) {
+                return new ResponseEntity<>("Unauthorized access to room", HttpStatus.FORBIDDEN);
+            }
             roomService.deleteRoom(id);
             return new ResponseEntity<>(HttpStatus.NO_CONTENT);
         } catch (EntityNotFoundException e) {
@@ -111,8 +143,13 @@ public class RoomController extends BaseController {
     @ApiResponse(responseCode = "404", description = "Room not found")
     @GetMapping("/{roomId}/devices")
     public ResponseEntity<?> getDevicesByRoomId(@Parameter(description = "ID of the room to retrieve devices for")
-                                                    @PathVariable Long roomId) {
+                                                    @PathVariable Long roomId, Authentication authentication) {
+        User user = ((CustomUserDetails) authentication.getPrincipal()).getUser();
         try {
+            Room room = roomService.getRoomById(roomId);
+            if (!homeService.userOwnsHome(user.getId(), room.getHome().getId())) {
+                return new ResponseEntity<>("Unauthorized access to room", HttpStatus.FORBIDDEN);
+            }
             List<Device> devices = roomService.getDevicesByRoomId(roomId);
             return new ResponseEntity<>(devices, HttpStatus.OK);
         } catch (EntityNotFoundException e) {
